last_updated_user_idカラムのデータ型を変更

+++ db/migrate/20200924020027_add_last_updated_user_to_pages.rb
@@ -2,6 +2,6 @@
 
 class AddLastUpdatedUserToPages < ActiveRecord::Migration[6.0]
   def change
-    add_column :pages, :last_updated_user_id, :bigint
+    add_column :pages, :last_updated_user_id, :integer
   end
 end
